move malloc to common folder
